% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_mcmcrs}
\alias{add_mcmcrs}
\title{Add mcmcr(s)}
\usage{
add_mcmcrs(x, x2)
}
\arguments{
\item{x}{An mcmcr or mcmcrs object.}

\item{x2}{An mcmcr or mcmcrs object.}
}
\value{
An object of class mcmcrs.
}
\description{
Add mcmcr(s)
}
