% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA.R
\name{anyNA.mcarray}
\alias{anyNA.mcarray}
\title{Any Missing Values}
\usage{
\method{anyNA}{mcarray}(x, recursive = FALSE)
}
\arguments{
\item{x}{The object to test.}

\item{recursive}{Unused.}
}
\description{
Returns a flag indicating whether there are any missing values.
}
\examples{
anyNA(mcmcr_example)
anyNA(mcmcr_example$beta)

}
