% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyNA.R
\name{anyNA}
\alias{anyNA}
\alias{anyNA.mcarray}
\alias{anyNA.mcmc}
\alias{anyNA.mcmc.list}
\alias{anyNA.mcmcarray}
\alias{anyNA.mcmcr}
\alias{anyNA.mcmcrs}
\title{Any Missing Values}
\usage{
\method{anyNA}{mcarray}(x, recursive = FALSE)

\method{anyNA}{mcmc}(x, recursive = FALSE)

\method{anyNA}{mcmc.list}(x, recursive = FALSE)

\method{anyNA}{mcmcarray}(x, recursive = FALSE)

\method{anyNA}{mcmcr}(x, recursive = FALSE)

\method{anyNA}{mcmcrs}(x, recursive = FALSE)
}
\arguments{
\item{x}{The object to test.}

\item{recursive}{Unused.}
}
\value{
A flag indicating whether there are any missing values.
}
\description{
Test whether there are any missing values.
}
\examples{
anyNA(mcmcr_example)
anyNA(mcmcr_example$beta)

}
\seealso{
\code{base::\link[base]{anyNA}}
}
