% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcmc.list}
\alias{as.mcmc.list}
\title{Coerce to an mcmc.list object}
\usage{
as.mcmc.list(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mcmc.list object
}

