% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmcrs.R
\name{as.mcmcrs}
\alias{as.mcmcrs}
\alias{as.mcmcrs.list}
\alias{as.mcmcrs.mcmcr}
\title{Convert to an mcmcrs object}
\usage{
as.mcmcrs(x, ...)

\method{as.mcmcrs}{list}(x, ...)

\method{as.mcmcrs}{mcmcr}(x, name = "mcmcr1", ...)
}
\arguments{
\item{x}{An MCMC object.}

\item{...}{Unused.}

\item{name}{A string specifying the element name.}
}
\value{
An mcmcrs object.
}
\description{
Converts an MCMC object to an \code{\link[=mcmcrs-object]{mcmcrs-object()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.mcmcrs(list)}: Convert a list of \verb{[mcmcr-object]s} to an mcmcrs object

\item \code{as.mcmcrs(mcmcr)}: Convert an \code{\link[=mcmcr-object]{mcmcr-object()}} to an mcmcrs object

}}
\examples{
as.mcmcrs(mcmcr::mcmcr_example)
}
