% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{as.term.character}
\alias{as.term.character}
\title{Coerce character vector to term vector}
\usage{
\method{as.term}{character}(x, ...)
}
\arguments{
\item{x}{The character vector to coerce}

\item{...}{Unused.}
}
\description{
Coerce character vector to term vector
}

