% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.term}
\alias{as.term}
\alias{as.character.term}
\alias{as.term.character}
\alias{as.term.mcmc}
\alias{as.term.mcmc.list}
\alias{as.term.mcmcarray}
\alias{as.term.mcmcr}
\title{Coerce to a term vector}
\usage{
as.term(x, ...)

\method{as.character}{term}(x, ...)

\method{as.term}{character}(x, ...)

\method{as.term}{mcmc}(x, ...)

\method{as.term}{mcmc.list}(x, ...)

\method{as.term}{mcmcarray}(x, ...)

\method{as.term}{mcmcr}(x, ...)
}
\arguments{
\item{x}{The object to coerce}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to a term vector.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Coerces term vector to a character vector

\item \code{character}: Coerces character vector to a term vector

\item \code{mcmc}: Coerces mcmc object to a term vector

\item \code{mcmc.list}: Coerces mcmc.list object to a term vector

\item \code{mcmcarray}: Coerces mcmcarray object to a term vector

\item \code{mcmcr}: Coerces mcmcr object to a term vector
}}

\examples{
as.term(mcmcr_example)
}
