% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains.mcarray}
\alias{bind_chains.mcarray}
\title{Combine MCMC objects by chains.}
\usage{
\method{bind_chains}{mcarray}(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined objects.
}
\description{
Combines two MCMC objects (with the same parameters and iterations) by chains.
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{fill_na}()},
\code{\link[universals]{split_chains}()}
}
