% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains.mcmc}
\alias{bind_chains.mcmc}
\title{Bind by Chains.}
\usage{
\method{bind_chains}{mcmc}(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined object.
}
\description{
Binds two MCMC objects (with the same parameters and iterations) by chains.
}
\examples{
bind_chains.foobar <- function(x, x2, ...) {
  NotYetImplemented()
  # replace with code to bind_chains for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{fill_na}()},
\code{\link[universals]{split_chains}()}
}
