% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains.mcmcarray}
\alias{bind_chains.mcmcarray}
\title{Combines objects by chains.}
\usage{
\method{bind_chains}{mcmcarray}(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines objects by chains.
}
