% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-dimensions.R
\name{bind_dimensions}
\alias{bind_dimensions}
\title{Combine two MCMC objects by dimensions}
\usage{
bind_dimensions(x, x2, along = NULL, ...)
}
\arguments{
\item{x}{An MCMC object.}

\item{x2}{a second MCMC object.}

\item{along}{A count (or NULL) indicating the parameter dimension to bind
along. For an \code{\link[=mcmcr-object]{mcmcr-object()}} it may also be a vector with one value per
parameter, in the order the parameters occur in \code{x}.}

\item{...}{Unused.}
}
\description{
Combines multiple MCMC objects (with the same parameters, chains and iterations)
by parameter dimensions.
}
\examples{
bind_dimensions(mcmcr_example, mcmcr_example)
}
\seealso{
\code{\link[universals:bind_chains]{universals::bind_chains()}}

Other bind:
\code{\link[=bind_dimensions_n]{bind_dimensions_n()}},
\code{\link[=bind_parameters]{bind_parameters()}}
}
\concept{bind}
