% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-iterations.R
\name{bind_iterations.mcmcr}
\alias{bind_iterations.mcmcr}
\title{Combines objects by iterations.}
\usage{
\method{bind_iterations}{mcmcr}(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines two or more mcmc objects by iterations
}
