% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-iterations.R
\name{bind_iterations}
\alias{bind_iterations}
\title{Combine two MCMC objects by iterations}
\usage{
bind_iterations(x, x2, ...)
}
\arguments{
\item{x}{an MCMC object}

\item{x2}{a second MCMC object}

\item{...}{unused}
}
\description{
Combines two MCMC objects (with the same parameters and chains) by iterations.
}
\examples{
bind_iterations(mcmcr_example, mcmcr_example)
}
