% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-parameters.R
\name{bind_parameters.mcmcr}
\alias{bind_parameters.mcmcr}
\title{Combines objects by parameters}
\usage{
\method{bind_parameters}{mcmcr}(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object}

\item{...}{Unused.}
}
\description{
Combines objects by parameters
}
