% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-parameters.R
\name{bind_parameters.mcmcrs}
\alias{bind_parameters.mcmcrs}
\title{Combines objects in mcmcrs by parameters.}
\usage{
\method{bind_parameters}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{...}{Unused.}
}
\description{
Ensures parameter names are distinct.
}
