% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-terms.R
\name{bind_terms.mcmc}
\alias{bind_terms.mcmc}
\title{Combines objects by terms}
\usage{
\method{bind_terms}{mcmc}(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines two or more mcmc objects by terms.
}
