% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-terms.R
\name{bind_terms}
\alias{bind_terms}
\title{Combines objects by terms}
\usage{
bind_terms(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{A second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines two mcmc objects by terms.
}
