% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mcmcarray}
\alias{check_mcmcarray}
\title{Check mcmcarray
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
check_mcmcarray(x, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (it if doesn't throw an error).
}
\description{
Check mcmcarray
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
check_mcmcarray(mcmcr::mcmcr_example$beta)
}
\seealso{
\code{\link[=check_mcmcr]{check_mcmcr()}}
}
