% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mcmcr}
\alias{check_mcmcr}
\title{Check mcmcr
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
check_mcmcr(x, sorted = FALSE, x_name = substitute(x), error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{sorted}{A flag specifying whether the parameters must be sorted.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (it if doesn't throw an error).
}
\description{
Check mcmcr
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
check_mcmcr(mcmcr::mcmcr_example)
}
\seealso{
\code{\link[=check_mcmcr]{check_mcmcr()}}
}
