% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_unused}
\alias{check_unused}
\title{Check Unused}
\usage{
check_unused(...)
}
\arguments{
\item{...}{The dots to check whether the are unused.}
}
\value{
An invisible NULL if unused or else throws an error.
}
\description{
Check Unused
}

