% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_mcmcr}
\alias{chk_mcmcr}
\alias{chk_mcmcarray}
\title{Check MCMC Objects}
\usage{
chk_mcmcarray(x, x_name = NULL)

chk_mcmcr(x, x_name = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
`NULL`, invisibly. Called for the side effect of throwing an error
  if the condition is not met.
}
\description{
Checks structure of MCMC objects.

`chk_mcmcarray`
checks if [mcmcarray-object()] object using

`is.array(x) && is.numeric(x)`

`chk_mcmcr`
checks if an [mcmcr-object()] while ignoring the class.
}
\details{
To check the class use \code{\link[chk]{chk_s3_class}(x, "class")}.
}
\section{Functions}{
\itemize{
\item \code{chk_mcmcarray}: Check mcmcarray Object

\item \code{chk_mcmcr}: Check mcmcr Object
}}

\examples{

# chk_mcmcarray
try(chk_mcmcarray(1))

# chk_mcmcr
chk_mcmcr(as.mcmcr(list(x = 1)))
try(chk_mcmcr(1))
}
\seealso{
[vld_mcmcr()]
}
