% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mcmcarray}
\alias{coef.mcmcarray}
\title{Coef mcmcarray}
\usage{
\method{coef}{mcmcarray}(object, conf_level = 0.95,
  estimate = stats::median, ...)
}
\arguments{
\item{object}{The object.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{A The function to use to calculate the estimate.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
Coefficients for an mcmcarray object.
}
\examples{
coef(mcmcr_example$beta)
}
