% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mcmcr}
\alias{coef.mcmcr}
\title{Coef mcmcr}
\usage{
\method{coef}{mcmcr}(object, conf_level = 0.95, estimate = stats::median,
  ...)
}
\arguments{
\item{object}{The object.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{A The function to use to calculate the estimate.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coefficient terms.
}
\description{
Coefficients for an mcmcr object.
}
\examples{
coef(mcmcr_example)
}
