% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse-chains.R
\name{collapse_chains}
\alias{collapse_chains}
\title{Collapse Chains}
\usage{
collapse_chains(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\description{
Collapses chains into a single chain.
}
