% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-chains.R
\name{combine_chains}
\alias{combine_chains}
\title{Combines objects by chains.}
\usage{
combine_chains(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines two or more mcmc objects by chains.
}

