% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-dimensions.R
\name{combine_dimensions}
\alias{combine_dimensions}
\title{Combines object dimensions}
\usage{
combine_dimensions(x, fun = mean, along = NULL, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{fun}{A function.}

\item{along}{A count (or NULL) indicating the parameter dimension(s) to bind along.}

\item{...}{Unused}
}
\value{
The object with reduced dimensions.
}
\description{
Combines mcmc object dimensions along along using fun.
}
