% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-values.R
\name{combine_values.mcmcr_data}
\alias{combine_values.mcmcr_data}
\title{Combine pair of mcmcr_data}
\usage{
\method{combine_values}{mcmcr_data}(x, x2, fun = mean, by = NULL,
  suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{x2}{A second mcmc object.}

\item{fun}{A function.}

\item{by}{A character vector of variables to join by}

\item{suffix}{If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to diambiguate them.}

\item{...}{Unused.}
}
\description{
Combine pair of mcmcr_data
}
