% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-values.R
\name{combine_values}
\alias{combine_values}
\title{Combines object values.}
\usage{
combine_values(x, x2, fun = mean, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{x2}{A second mcmc object.}

\item{fun}{A function.}

\item{...}{Unused.}
}
\description{
Combines mcmc objects by values using fun.
}
