% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-values.R
\name{combine_values_list}
\alias{combine_values_list}
\title{Combine list of mcmc samples}
\usage{
combine_values_list(x, fun = mean, by = NULL, suffix = c(".x", ".y"))
}
\arguments{
\item{x}{An mcmc object.}

\item{fun}{A function.}

\item{by}{A character vector of variables to join by if combining mcmcr_data objects.}

\item{suffix}{If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to diambiguate them if combininig mcmcr_data objects.}
}
\description{
Combine list of mcmc samples
}
