% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged.mcmcr}
\alias{converged.mcmcr}
\title{Is Converged}
\usage{
\method{converged}{mcmcr}(x, rhat = 1.1, efs = 0.33, ...)
}
\arguments{
\item{x}{The object to test.}

\item{rhat}{A number specifying the rhat threshold.}

\item{efs}{A number specifying the minimum effective sampling rate.}

\item{...}{Unused}
}
\description{
Is Converged
}
