% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged.mcmcrs}
\alias{converged.mcmcrs}
\title{Is Converged}
\usage{
\method{converged}{mcmcrs}(x, rhat = 1.1, ...)
}
\arguments{
\item{x}{The object to test.}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused}
}
\description{
Is Converged
}
