% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Convergence}
\usage{
convergence(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the convergence value.
}
\description{
Get R-hat value(s) for JAGS objects
}

