% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-rhat.R
\name{cross_converged.mcmcarray}
\alias{cross_converged.mcmcarray}
\title{Is Converged}
\usage{
\method{cross_converged}{mcmcarray}(x, x2, rhat = 1.1, ...)
}
\arguments{
\item{x}{The first object to test.}

\item{x2}{The second object to test against.}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused}
}
\description{
Is Converged
}
