% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-rhat.R
\name{cross_converged}
\alias{cross_converged}
\title{Have Two Objects Converged}
\usage{
cross_converged(x, x2, ...)
}
\arguments{
\item{x}{The first object to test.}

\item{x2}{The second object to test against.}

\item{...}{Unused}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether two objects have converged on the same answers.
}
