% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-rhat.R
\name{cross_rhat}
\alias{cross_rhat}
\title{Convergence}
\usage{
cross_rhat(x, x2, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{x2}{A second mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the rhat value.
}
\description{
Get R-hat value(s) for convergence between two mcmc objects
}
