% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive.mcmcr}
\alias{derive.mcmcr}
\title{derive}
\usage{
\method{derive}{mcmcr}(object, expr, values = list(), monitor = ".*",
  parallel = FALSE, quick = FALSE, ...)
}
\arguments{
\item{object}{The mcmcr object.}

\item{expr}{A string of the R expression to evaluate.}

\item{values}{A lists of the values to use.}

\item{monitor}{A regular expression specifying the new variables to monitor.}

\item{parallel}{A flag indicating whether to derive samples in parallel using foreach backend.}

\item{quick}{A flag indicating whether to quickly calculate deriveions.}

\item{...}{Unused.}
}
\value{
An mcmcr object of the monitored parameters.
}
\description{
Calculate dervived parameters.
}
