% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{derive}
\alias{derive}
\title{Derived parameters}
\usage{
derive(object, ...)
}
\arguments{
\item{object}{The object to calculate the number of derived parameters for.}

\item{...}{Not used.}
}
\description{
Derived parameters
}

