% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dims}
\alias{dims}
\title{Dimensions}
\usage{
dims(x)
}
\arguments{
\item{x}{The object.}
}
\value{
An integer vector of the dimensions.
}
\description{
Dimensions
}
\examples{
dims(character())
dims(2:3)
dims(matrix(1:6, nrow = 2))
}
