% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{dims_term}
\alias{dims_term}
\title{Get dims for term object}
\usage{
dims_term(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused}
}
\value{
A list of the dims.
}
\description{
Get dims for term object
}

