% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp.R
\name{esp}
\alias{esp}
\title{Effective Sampling Proportion}
\usage{
esp(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the esp value.
}
\description{
Effective Sampling Proportion
}
