% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr.R
\name{esr.mcmc}
\alias{esr.mcmc}
\title{Effective Sampling Rate}
\usage{
\method{esr}{mcmc}(x, by = "all", as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{...}{Other arguments passed to methods.}
}
\value{
The esr value(s).
}
\description{
Calculates the effective sampling rate (\code{esr}) based on the formula
\deqn{\frac{1}{1 + 2 \sum_{k = 1}^\infty\rho_k(\theta)}}
in Brooks et al. (2011).
The infinite sum is truncated at
lag \eqn{k} when \eqn{\rho_{k+1}(\theta) < 0}.
}
\references{
Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.
}
