% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mcarray}
\alias{estimates.mcarray}
\title{Estimates}
\usage{
\method{estimates}{mcarray}(x, fun = median, as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{as_df}{A flag indicating whether to return the results as a
data frame versus a named list.}

\item{...}{Optional arguments to fun.}
}
\value{
A named list or data frame.
}
\description{
Calculates the estimates for an MCMC object.
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{bind_iterations}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{split_chains}()}
}
