% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates.mcmc.list}
\alias{estimates.mcmc.list}
\title{Estimates}
\usage{
\method{estimates}{mcmc.list}(x, fun = median, as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list or data frame.
}
\description{
Calculates the estimates for an MCMC object.
}
\examples{
estimates.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get estimates for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{split_chains}()}
}
