% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_mcmcr_data}
\alias{filter_mcmcr_data}
\title{Filter mcmcr data}
\usage{
filter_mcmcr_data(.x, ...)
}
\arguments{
\item{.x}{The mcmcr_object to filter}

\item{...}{Logical predicates. Multiple conditions are combined with &.}
}
\description{
Filter mcmcr data
}
