% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_mcmcr_data_}
\alias{filter_mcmcr_data_}
\title{Filter mcmcr data}
\usage{
filter_mcmcr_data_(.x, ..., .dots)
}
\arguments{
\item{.x}{The mcmcr_object to filter}

\item{...}{Logical predicates. Multiple conditions are combined with &.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}
}
\description{
Filter mcmcr data
}
