% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_by_.mcmcr_data}
\alias{group_by_.mcmcr_data}
\title{Group mcmcr data}
\usage{
\method{group_by_}{mcmcr_data}(.data, ..., .dots, add = FALSE)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Variables to group by. All tbls accept variable names.
Some tbls will accept functions of variables. Duplicated groups
will be silently dropped.}

\item{.dots}{Used to work around non-standard evaluation.
See \code{vignette("nse")} for details.}

\item{add}{When \code{add = FALSE}, the default, \code{group_by()} will
override existing groups. To add to the existing groups, use
\code{add = TRUE}.}
}
\description{
Group mcmcr data
}
