% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_by_mcmcr_data}
\alias{group_by_mcmcr_data}
\title{Group mcmcr data}
\usage{
group_by_mcmcr_data(.x, ..., add = FALSE)
}
\arguments{
\item{.x}{The mcmcr_object to group}

\item{...}{variables to group by. All tbls accept variable names, some will also accept functions of variables. Duplicated groups will be silently dropped.}

\item{add}{By default, when add = FALSE, group_by will override existing groups. To instead add to the existing groups, use add = TRUE}
}
\description{
Group mcmcr data
}
