% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_by_mcmcr_data_}
\alias{group_by_mcmcr_data_}
\title{Group mcmcr data}
\usage{
group_by_mcmcr_data_(.x, ..., .dots, add = FALSE)
}
\arguments{
\item{.x}{The mcmcr_object to group}

\item{...}{variables to group by. All tbls accept variable names, some will also accept functions of variables. Duplicated groups will be silently dropped.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}

\item{add}{By default, when add = FALSE, group_by will override existing groups. To instead add to the existing groups, use add = TRUE}
}
\description{
Group mcmcr data
}
