% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{ilogit<-}
\alias{ilogit<-}
\title{Inverse Logistic Transformation}
\usage{
ilogit(x) <- value
}
\arguments{
\item{x}{The object to replace.}

\item{value}{The numeric vector to transform}
}
\description{
Inverse Logistic Transformation
}
\examples{
x <- 1
ilogit(x) <- 0.5
x
}
