% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{ilogit}
\alias{ilogit}
\title{Inverse Logistic Transformation}
\usage{
ilogit(x)
}
\arguments{
\item{x}{The numeric vector to transform.}
}
\description{
Inverse Logistic Transformation
}
\examples{
ilogit(c(0.25,0.5,0.75))
}
