% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmc.list}
\alias{is.mcmc.list}
\title{Is an mcmc.list object}
\usage{
is.mcmc.list(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmc.list'
}
\examples{
is.mcmc.list(coda::as.mcmc.list(mcmcr_example))
}
