% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmc}
\alias{is.mcmc}
\title{Is an mcmc object}
\usage{
is.mcmc(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmc'
}
\examples{
is.mcmc(coda::as.mcmc.list(mcmcr_example)[[1]])
}
