% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmcr}
\alias{is.mcmcr}
\title{Is an mcmcr object}
\usage{
is.mcmcr(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmcr'
}
