% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmcr_data}
\alias{is.mcmcr_data}
\title{Is an mcmcr data object}
\usage{
is.mcmcr_data(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmcr_data'
}
