% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmcrs}
\alias{is.mcmcrs}
\title{Is mcmcrs Object}
\usage{
is.mcmcrs(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an object is an \code{\link{mcmcrs-object}}.
}
\examples{
is.mcmcrs(mcmcrs(mcmcr_example))
}
