% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{is_converged.default}
\alias{is_converged.default}
\title{Is Converged}
\usage{
is_converged.default(x, rhat = 1.1, ...)
}
\arguments{
\item{x}{The object to test.}

\item{rhat}{A number specifying the rhat threshold.}

\item{...}{Unused}
}
\description{
Is Converged
}

