% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_mcarrays}
\alias{list_mcarrays}
\title{List of McArrays}
\format{An object of class \code{list} of length 3.}
\usage{
list_mcarrays
}
\description{
A list of objects of class mcarray.
}
\keyword{datasets}

