% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{logit<-}
\alias{logit<-}
\title{Logistic Transformation}
\usage{
logit(x) <- value
}
\arguments{
\item{x}{The object to replace.}

\item{value}{The numeric vector to transform}
}
\description{
Logistic Transformation
}
\examples{
x <- 1
logit(x) <- 0.5
x
}
