% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-converged.R
\name{mcmc_converged}
\alias{mcmc_converged}
\alias{mcmc_converged.default}
\title{MCMC Object Converged}
\usage{
mcmc_converged(x, rhat = 1.1, esr = 0.33, by = "all", as_df = FALSE,
  ...)

\method{mcmc_converged}{default}(x, rhat = 1.1, esr = 0.33, by = "all",
  as_df = FALSE, ...)
}
\arguments{
\item{x}{An MCMC object.}

\item{rhat}{The maximum rhat value.}

\item{esr}{The minimum effective sampling rate.}

\item{by}{A string indicating whether to determine by "term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the results as a (tbl) data frame.}

\item{...}{Unused}
}
\description{
Test whether an MCMC object has converged.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Test whether an MCMC object has converged
}}

\examples{
mcmc_converged(mcmcr_example)
}
