% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-object.R
\name{mcmcr-object}
\alias{mcmcr-object}
\title{mcmcr}
\description{
An \code{mcmcr} object stores multiple uniquely named \code{\link{mcmcarray-object}}
objects with the same number of chains and iterations.
}
\details{
\code{mcmcr} objects allow a set of dimensionality preserving parameters
to be manipulated and queried as a whole.
}
\examples{
mcmcr_example
}
