% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-data.R
\name{mcmcr_data}
\alias{mcmcr_data}
\title{MCMC Data}
\usage{
mcmcr_data(mcmcr, data)
}
\arguments{
\item{mcmcr}{An mcmcr object}

\item{data}{A data.frame}
}
\value{
An object of class \code{mcmc_datar}
}
\description{
Combines an mcmcr object with samples from a single vector with a data frame with the same number of rows.
The resultant object can be filtered or summarized or the coef calculated.
}
