% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcrs-object.R
\name{mcmcrs-object}
\alias{mcmcrs-object}
\alias{mcmcrs_object}
\title{mcmcrs}
\description{
An \code{mcmcrs} object stores multiple \code{\link{mcmcr-object}}s
with the same parameters and the same number of chains and iterations.
}
\details{
\code{mcmcrs} objects allow the results of multiple analyses
using the same model to be manipulated and queried as a whole.
}
\examples{
mcmcrs(mcmcr_example, mcmcr_example)
}
