% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcrs.R
\name{mcmcrs}
\alias{mcmcrs}
\title{Create mcmcrs}
\usage{
mcmcrs(...)
}
\arguments{
\item{...}{Objects of class mcmcr.}
}
\value{
An object of class mcmcrs
}
\description{
Creates an \code{\link[=mcmcrs-object]{mcmcrs-object()}} from multiple \code{link{mcmcr-object}}s.
}
\examples{
mcmcrs(mcmcr_example, mcmcr_example)
}
