% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MetropHastings.R
\name{mh}
\alias{mh}
\title{Metropolis-Hastings MCMC sampling}
\usage{
mh(dat, start, LL, priors, N, giveup = 100)
}
\arguments{
\item{dat}{dataset to be based to the likelihood function}

\item{start}{a vector of starting values}

\item{LL}{a likelihood function that accepts the parameters and data}

\item{priors}{functions that give the prior probability
for the parameter}

\item{N}{total number of sampling rounds (one rounds goes through
all parameters once).}

\item{giveup}{after how many rounds of no new accepted values
does the algorithm give up and assume convergence?}
}
\description{
Simple protocol for a MCMC sampling algorithm.
Given a set of priors, and a likelihood function, the algorithm
samples new values for N parameters from a standard normal with
a jump variance that is automatically adjusted. The algoritm
accepts new proposal proportional to a random fraction.
}
\author{
Marco D. Visser
}

