% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains.mcarray}
\alias{nchains.mcarray}
\title{Number of MCMC Chains}
\usage{
\method{nchains}{mcarray}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of MCMC chains.
}
\description{
Gets the number of chains
of a Monte Carlo Markov Chain (MCMC) object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
