% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains.mcmc.list}
\alias{nchains.mcmc.list}
\title{Number of Chains}
\usage{
\method{nchains}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of chains.
}
\description{
Gets the number of chains
of an MCMC object.
}
\examples{
\dontshow{if (requireNamespace("nlist", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(nlist)

nchains(nlist(x = 1:2))
nlists <- nlists(nlist(x = c(2, 9)), nlist(x = c(1, 7)))
nchains(nlists)
nchains(split_chains(nlists))
\dontshow{\}) # examplesIf}
}
\seealso{
Other MCMC dimensions: 
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
