% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains.mcmcarray}
\alias{nchains.mcmcarray}
\title{Number of Chains}
\usage{
\method{nchains}{mcmcarray}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of chains.
}
\description{
Gets the number of chains
of an MCMC object.
}
\examples{
nchains.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get nchains for an object of class 'foobar'
}
}
\seealso{
Other MCMC dimensions: 
\code{\link[universals]{niters}()},
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
