% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ndims}
\alias{ndims}
\title{Number of dimensions}
\usage{
ndims(x)
}
\arguments{
\item{x}{The object to get the number of dimensions of.}
}
\value{
A count of the number of dimensions
}
\description{
Number of dimensions
}
