% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters.mcmc}
\alias{niters.mcmc}
\title{Number of Iterations}
\usage{
\method{niters}{mcmc}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of iterations.
}
\description{
Gets the number of iterations (in a chain)
of an MCMC object.
}
\seealso{
Other MCMC dimensions: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
