% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters.mcmcarray}
\alias{niters.mcmcarray}
\title{Number of MCMC Iterations}
\usage{
\method{niters}{mcmcarray}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of MCMC iterations.
}
\description{
Gets the number of iterations (in a chain)
of a Monte Carlo Markov Chain (MCMC) object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}
}
