% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters}
\alias{niters}
\alias{niters.mcarray}
\alias{niters.mcmc}
\alias{niters.mcmc.list}
\alias{niters.mcmcarray}
\alias{niters.mcmcr}
\alias{niters.mcmcrs}
\title{Number of MCMC samples}
\usage{
niters(x, ...)

\method{niters}{mcarray}(x, ...)

\method{niters}{mcmc}(x, ...)

\method{niters}{mcmc.list}(x, ...)

\method{niters}{mcmcarray}(x, ...)

\method{niters}{mcmcr}(x, ...)

\method{niters}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count indicating the number of MCMC iterations.
}
\description{
Gets the number of MCMC iterations (in a chain).
}
\section{Methods (by class)}{
\itemize{
\item \code{mcarray}: Number of MCMC iterations for an mcarray object

\item \code{mcmc}: Number of MCMC iterations for an mcmc object

\item \code{mcmc.list}: Number of MCMC iterations for an mcmc.list object

\item \code{mcmcarray}: Number of MCMC iterations for an mcmcarray object

\item \code{mcmcr}: Number of MCMC iterations for an mcmcr object

\item \code{mcmcrs}: Number of MCMC iterations for an mcmcrs object
}}

\examples{
niters(mcmcr_example)
}
