% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars.mcarray}
\alias{npars.mcarray}
\title{Number of Parameters}
\usage{
\method{npars}{mcarray}(x, scalar = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of parameters.
}
\description{
Gets the number of parameters of an object.
}
\examples{
npars.foobar <- function(x, ...) {
  length(pars(x, ...))
}
}
\seealso{
\code{\link[universals:pars]{pars()}}

Other MCMC dimensions: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}

Other parameters: 
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}
}
