% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars.mcmc.list}
\alias{npars.mcmc.list}
\title{Number of Parameters}
\usage{
\method{npars}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of parameters.
}
\description{
Gets the number of parameters of an object.
}
\seealso{
\code{\link[universals:pars]{pars()}}

Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}

Other {parameters}: 
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{pars_scalar}()},
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}
}
