% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\alias{npars.default}
\alias{npars.mcarray}
\alias{npars.mcmcarray}
\alias{npars.mcmcr}
\alias{npars.mcmcrs}
\title{Number of Parameters}
\usage{
npars(x, ...)

\method{npars}{default}(x, ...)

\method{npars}{mcarray}(x, ...)

\method{npars}{mcmcarray}(x, ...)

\method{npars}{mcmcr}(x, ...)

\method{npars}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A count of the number of parameters.
}
\description{
Gets the number of parameters for an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Number of parameters for an object

\item \code{mcarray}: Number of parameters for an mcarray object

\item \code{mcmcarray}: Number of parameters for an mcmcarray object

\item \code{mcmcr}: Number of parameters for an mcmcr object

\item \code{mcmcrs}: Number of parameters for an mcmcrs object
}}

\examples{
npars(mcmcr_example)
}
