% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims.mcmcr}
\alias{npdims.mcmcr}
\title{Number of Parameter Dimensions}
\usage{
\method{npdims}{mcmcr}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named integer vector of the number of dimensions of each parameter.
}
\description{
Gets the number of the dimensions of each parameter of an object.
}
\examples{
npdims.foobar <- function(x, ...) {
  vapply(pdims(x, ...), length, 1L)
}
}
\seealso{
Other dimensions: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{pdims}()}
}
