% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims}
\alias{npdims}
\title{Number of Parameter Dimensions}
\usage{
npdims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the number parameter dimensions of an object.
}
\examples{
pdims(mcmcr_example)
}
