% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsamples}
\alias{nsamples}
\title{Number of MCMC Samples}
\usage{
nsamples(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A count of the total number of samples.
}
\description{
The product of the number of chains and number of iterations in each chain.
}
