% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsams}
\alias{nsams}
\title{Number of MCMC Samples}
\usage{
nsams(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A count of the total number of samples.
}
\description{
The product of the number of simulations and the number of terms.
}
\examples{
nsams(mcmcr_example)
}
