% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsims.R
\name{nsims.default}
\alias{nsims.default}
\title{Number of Simulations}
\usage{
\method{nsims}{default}(x, ...)
}
\arguments{
\item{x}{The MCMC object}

\item{...}{Unused.}
}
\value{
A count of the number of simulations.
}
\description{
Gets the number of simulations of an MCMC object.
}
\examples{
nlist::nsims(mcmcr::mcmcr_example)
}
