% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms.mcmcrs}
\alias{nterms.mcmcrs}
\title{Number of Terms}
\usage{
\method{nterms}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A integer scalar of the number of terms.
}
\description{
Gets the number of terms
of an MCMC object.
}
\seealso{
Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()}

Other {terms}: 
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{pars_terms}()}
}
