% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\alias{nterms.default}
\alias{nterms.mcmcarray}
\alias{nterms.mcmcr}
\alias{nterms.mcmcrs}
\title{Number of Terms}
\usage{
nterms(x, ...)

\method{nterms}{default}(x, ...)

\method{nterms}{mcmcarray}(x, ...)

\method{nterms}{mcmcr}(x, ...)

\method{nterms}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms for an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Number of terms for an object

\item \code{mcmcarray}: Number of terms for an mcmcarray object

\item \code{mcmcr}: Number of terms for an mcmcr object

\item \code{mcmcrs}: Number of terms for an mcmcrs object
}}

\examples{
nterms(mcmcr_example)
}
