% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-dims.R
\name{parameter_dims}
\alias{parameter_dims}
\title{Parameter Dimensions}
\usage{
parameter_dims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the parameter dimensions of an object.
}
