% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-ndims.R
\name{parameter_ndims}
\alias{parameter_ndims}
\title{Number of Parameter Dimensions}
\usage{
parameter_ndims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets an integer vector of the number of dimensions of the parameters of an object.
}
