% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-pars.R
\name{parameters<-}
\alias{parameters<-}
\title{Set Parameters}
\usage{
parameters(x) <- value
}
\arguments{
\item{x}{An object.}

\item{value}{A character vector of the new parameter names.}
}
\value{
The modified object.
}
\description{
Sets an object's parameter names.
}
\details{
\code{value} must be a unique character vector of the same length as the
object's parameters.

\lifecycle{soft-deprecated} for pars<-()
}
