% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_parameters.R
\name{parameters.mcmc}
\alias{parameters.mcmc}
\title{Parameters}
\usage{
\method{parameters}{mcmc}(x, scalar_only = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{scalar_only}{A flag indicating whether to only get scalar parameters.}

\item{...}{Not used.}
}
\description{
Parameters
}
