% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters.mcmc}
\alias{parameters.mcmc.list}
\alias{parameters.mcmcr}
\alias{parameters.mcmcrs}
\title{Parameter Names}
\usage{
parameters(x, ...)

\method{parameters}{mcmc}(x, scalar = NULL, terms = FALSE, ...)

\method{parameters}{mcmc.list}(x, scalar = NULL, terms = FALSE, ...)

\method{parameters}{mcmcr}(x, scalar = NULL, terms = FALSE, ...)

\method{parameters}{mcmcrs}(x, scalar = NULL, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\details{
\lifecycle{soft-deprecated} for pars()
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmc}: Parameters mcmc

\item \code{mcmc.list}: Parameters mcmc.list

\item \code{mcmcr}: Parameters mcmcr

\item \code{mcmcrs}: Parameters mcmcrs
}}

