% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{parameters_term}
\alias{parameters_term}
\title{Get parameters for term object}
\usage{
parameters_term(x)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused}
}
\value{
A vector of the parameter names.
}
\description{
Get parameters for term object
}
