% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.mcmc}
\alias{pars.mcmc}
\title{Parameter Names}
\usage{
\method{pars}{mcmc}(x, scalar = NA, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to only include scalar terms.}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
pars.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get pars for an object of class 'foobar'
}
}
\seealso{
Other parameters: 
\code{\link[universals]{npars_scalar}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{pars_scalar}()},
\code{\link[universals]{set_pars}()}
}
