% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims.mcmcr}
\alias{pdims.mcmcr}
\title{Parameter Dimensions}
\usage{
\method{pdims}{mcmcr}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of integer vectors of the dimensions of each parameter.
}
\description{
Gets the dimensions of each parameter of an object.
}
\examples{
pdims.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get pdims for an object of class 'foobar'
}
}
\seealso{
Other dimensions: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()}
}
