% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R, R/pdims.R
\name{npdims.default}
\alias{npdims.default}
\alias{npdims.mcmc.list}
\alias{npdims.mcmcr}
\alias{pdims}
\alias{pdims.term}
\alias{pdims.mcarray}
\alias{pdims.mcmc}
\alias{pdims.mcmc.list}
\alias{pdims.mcmcarray}
\alias{pdims.mcmcr}
\title{Parameter Dimensions}
\usage{
\method{npdims}{default}(x, ...)

\method{npdims}{mcmc.list}(x, ...)

\method{npdims}{mcmcr}(x, ...)

pdims(x, ...)

\method{pdims}{term}(x, ...)

\method{pdims}{mcarray}(x, ...)

\method{pdims}{mcmc}(x, ...)

\method{pdims}{mcmc.list}(x, ...)

\method{pdims}{mcmcarray}(x, ...)

\method{pdims}{mcmcr}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the parameter dimensions of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Parameter dimensions for a term vector

\item \code{mcmc.list}: Parameter dimensions for an mcmcr object

\item \code{mcmcr}: Parameter dimensions for an mcmcr object

\item \code{term}: Parameter dimensions for a term vector

\item \code{mcarray}: Parameter dimensions for an mcarray object

\item \code{mcmc}: Parameter dimensions for an mcmc object

\item \code{mcmc.list}: Parameter dimensions for an mcmc.list object

\item \code{mcmcarray}: Parameter dimensions for an mcmcarray object

\item \code{mcmcr}: Parameter dimensions for an mcmcr object
}}

\examples{
pdims(mcmcr_example)
}
