% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{pow}
\alias{pow}
\title{Power}
\usage{
pow(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{A numeric vector of the power term.}
}
\description{
R equivalent to the C++  function.
}
\examples{
pow(10,2)
}
