% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mcmcr}
\alias{predict.mcmcr}
\title{Predict}
\usage{
\method{predict}{mcmcr}(object, values, expr, parameters = ".*", ...)
}
\arguments{
\item{object}{The mcmcr object.}

\item{values}{A lists of the values to use.}

\item{expr}{A string of the R expression}

\item{parameters}{A regular expression indication the parameter to monitor.}

\item{...}{Unused.}
}
\value{
An mcmcr object of the monitored parameters.
}
\description{
Calculate predictions.
}

