% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pvalue}
\alias{pvalue}
\title{P-Value}
\usage{
pvalue(x)
}
\arguments{
\item{x}{A vector of MCMC value}
}
\value{
A number indicating the p-value.
}
\description{
Calculates the p-value.
}
\examples{
pvalue(as.numeric(0:10))
}
