% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat.mcmc.list}
\alias{rhat.mcmc.list}
\title{R-hat}
\usage{
\method{rhat}{mcmc.list}(x, by = "all", as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{...}{Other arguments passed to methods.}
}
\value{
The rhat value(s).
}
\description{
Calculates the uncorrected, unfolded, univariate,
split R-hat (potential scale reduction factor) value(s).
}
\references{
Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
}
