% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat.mcmcr}
\alias{rhat.mcmcr}
\title{R-hat}
\usage{
\method{rhat}{mcmcr}(x, by = "all", as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{...}{Other arguments passed to methods.}
}
\value{
A number >= 1 indicating the rhat value.
}
\description{
Calculates the uncorrected, unfolded, univariate,
split R-hat (potential scale reduction factor) value.
}
\examples{
rhat(mcmcr_example)
rhat(mcmcr_example, by = "parameter")
rhat(mcmcr_example, by = "term")
rhat(mcmcr_example, by = "term", as_df = TRUE)
}
\references{
Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
}
\seealso{
Other {convergence}: 
\code{\link[universals]{converged}()},
\code{\link[universals]{esr}()}
}
