% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat.mcmcrs}
\alias{rhat.mcmcrs}
\title{R-hat}
\usage{
\method{rhat}{mcmcrs}(x, by = "all", as_df = FALSE, bound = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{bound}{flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.}

\item{...}{Other arguments passed to methods.}
}
\value{
The rhat value(s).
}
\description{
Calculates the uncorrected, unfolded, univariate,
split R-hat (potential scale reduction factor) value(s).
}
\examples{
rhat(mcmcrs(mcmcr_example, mcmcr_example))
rhat(mcmcrs(mcmcr_example, mcmcr_example), bound = TRUE)
}
\references{
Gelman, A., and Rubin, D.B. 1992. Inference from Iterative Simulation Using Multiple Sequences. Statistical Science 7(4): 457–472.
}
