% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RejectionSampling.R
\name{rs}
\alias{rs}
\title{rejection sampling}
\usage{
rs(dat, start, LL, priors, N, giveup = 100)
}
\arguments{
\item{dat}{dataset to be based to the likelihood function}

\item{start}{a vector of starting values}

\item{LL}{a likelihood function that accepts the parameters and data}

\item{priors}{a function that randomly samples parameters from
their prior distributions}

\item{N}{total number of sampling rounds (one rounds goes through
all parameters once).}

\item{giveup}{after how many rounds of no new accepted values
does the algorithm give up and assume convergence?}
}
\description{
Protocol for a simple rejection sampling algorithm.
Given a set of priors to sample N parameters from
the algoritm accepts new proposal proportional to a random
fraction. The sequence of parameters are randomly reversed.
}
\author{
Marco D. Visser
}

