% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_parameters.R
\name{scalar_parameters}
\alias{scalar_parameters}
\title{Scalar Parameters}
\usage{
scalar_parameters(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the scalar parameter names for an object.
}
