% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_parameters}
\alias{set_parameters}
\title{Set Parameters}
\usage{
set_parameters(x, parameters)
}
\arguments{
\item{x}{The object}

\item{parameters}{A character vector the new parameter names.}
}
\value{
A copy of the updated object
}
\description{
Set Parameters
}
