% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_.mcmcr_data}
\alias{slice_.mcmcr_data}
\title{Slice mcmcr data}
\usage{
\method{slice_}{mcmcr_data}(.data, ..., .dots)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Pair/values of expressions coercible to lazy objects.}

\item{.dots}{Pair/values of expressions coercible to lazy objects.}
}
\description{
Slice mcmcr data
}
