% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort}
\alias{sort}
\alias{sort.mcmc}
\alias{sort.mcmc.list}
\alias{sort.mcmcr}
\alias{sort.mcmcrs}
\title{Sort an MCMC Object}
\usage{
\method{sort}{mcmc}(x, ...)

\method{sort}{mcmc.list}(x, ...)

\method{sort}{mcmcr}(x, ...)

\method{sort}{mcmcrs}(x, ...)
}
\arguments{
\item{x}{The MCMC object to sort}

\item{...}{Unused}
}
\description{
Sorts an MCMC object by its parameter names.
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmc}: Sort an mcmc object

\item \code{mcmc.list}: Sort an mcmc.list object

\item \code{mcmcr}: Sort an mcmcr object

\item \code{mcmcrs}: Sort an mcmcrs object
}}

\examples{
parameters(mcmcr_example)
mcmcr_example <- subset(mcmcr_example, parameters = c("beta", "alpha"))
parameters(mcmcr_example)
mcmcr_example <- sort(mcmcr_example)
parameters(mcmcr_example)
}
