% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains.mcmcarray}
\alias{split_chains.mcmcarray}
\title{Split Chains}
\usage{
\method{split_chains}{mcmcarray}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Splits each of an MCMC object's chains in half
to double the number chains
and halve the number of iterations.
}
\seealso{
Other {MCMC manipulations}: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{fill_all}()},
\code{\link[universals]{fill_na}()}
}
