% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.mcmcr}
\alias{subset.mcmcr}
\title{Subset mcmc object}
\usage{
\method{subset}{mcmcr}(x, chains = NULL, iterations = NULL,
  parameters = NULL, ...)
}
\arguments{
\item{x}{The object to subset}

\item{chains}{An integer vector of the chains to subset by.}

\item{iterations}{An integer vector of the iterations to subset by.}

\item{parameters}{A character vector of  the parameters to subset by.}

\item{...}{Unused.}
}
\description{
Subset mcmc object
}
