% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_mcmcr_data}
\alias{summarise_mcmcr_data}
\title{Summarise mcmcr data}
\usage{
summarise_mcmcr_data(.x, ..., .fun = mean)
}
\arguments{
\item{.x}{The mcmcr_object to filter}

\item{...}{Name-value pairs of summary functions like min(), mean(), max() etc.}

\item{.fun}{The function to summarise the samples by.}
}
\description{
Summarise mcmcr data
}
