% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarize_mcmcr_data_}
\alias{summarize_mcmcr_data_}
\title{Summarise mcmcr data}
\usage{
summarize_mcmcr_data_(.x, ..., .dots, .fun = mean)
}
\arguments{
\item{.x}{The mcmcr_object to filter}

\item{...}{Name-value pairs of summary functions like min(), mean(), max() etc.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}

\item{.fun}{The function to summarise the samples by.}
}
\description{
Summarise mcmcr data
}
