% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdims.R
\name{tdims}
\alias{tdims}
\alias{tdims.default}
\alias{tdims.term}
\title{Term Dimensions}
\usage{
tdims(x, ...)

\method{tdims}{default}(x, ...)

\method{tdims}{term}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the term dimensions of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Term dimensions for an object

\item \code{term}: Term dimensions for a term vector
}}

\examples{
tdims(mcmcr_example)
}
