% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms}
\alias{terms}
\alias{terms.mcmc}
\alias{terms.mcmc.list}
\alias{terms.mcmcarray}
\alias{terms.mcmcr}
\title{MCMC Object Terms}
\usage{
\method{terms}{mcmc}(x, ...)

\method{terms}{mcmc.list}(x, ...)

\method{terms}{mcmcarray}(x, ...)

\method{terms}{mcmcr}(x, ...)
}
\arguments{
\item{x}{A MCMC object.}

\item{...}{Unused}
}
\value{
A term vector.
}
\description{
Gets terms for an MCMC object.
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmc}: Get terms for an mcmc object

\item \code{mcmc.list}: Get terms for an mcmc.list object

\item \code{mcmcarray}: Get terms for an mcmcarray object

\item \code{mcmcr}: Get terms for an mcmcr object
}}

\examples{
terms(mcmcr_example)
}
\seealso{
\code{stats::\link[stats]{terms}}
}
