% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\alias{thin.mcmcarray}
\alias{thin.mcmcr}
\alias{thin.mcmcrs}
\title{Thin MCMC Samples}
\usage{
\method{thin}{mcmcarray}(x, nthin = 1L, ...)

\method{thin}{mcmcr}(x, nthin = 1L, ...)

\method{thin}{mcmcrs}(x, nthin = 1L, ...)
}
\arguments{
\item{x}{An MCMC object}

\item{nthin}{A count of the thinning rate.}

\item{...}{Unused}
}
\value{
The thinned object.
}
\description{
Thin an MCMC objects samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmcarray}: Thin MCMC samples for an mcmcarray object

\item \code{mcmcr}: Thin MCMC samples for an mcmcr object

\item \code{mcmcrs}: Thin MCMC samples for an mcmcrs object
}}

\examples{
thin(mcmcr_example, nthin = 10L)
}
\seealso{
\code{coda::\link[coda]{thin}}
}
