% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{ungroup.mcmcr_data}
\alias{ungroup.mcmcr_data}
\title{Ungroup mcmcr data}
\usage{
\method{ungroup}{mcmcr_data}(x, ...)
}
\arguments{
\item{x}{An mcmcr_data object}

\item{...}{Variables to group by. All tbls accept variable names,
some will also accept functions of variables. Duplicated groups
will be silently dropped.}
}
\description{
Ungroup mcmcr data
}
