% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{ungroup_mcmcr_data}
\alias{ungroup_mcmcr_data}
\title{Ungroup mcmcr data}
\usage{
ungroup_mcmcr_data(.x, ...)
}
\arguments{
\item{.x}{The mcmcr_object to group}

\item{...}{Additional arguments that maybe used by methods.}
}
\description{
Ungroup mcmcr data
}
