% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_mcmcr}
\alias{vld_mcmcr}
\alias{vld_mcmcarray}
\alias{vld_mcmcrs}
\title{Validate MCMC Objects}
\usage{
vld_mcmcarray(x)

vld_mcmcr(x)

vld_mcmcrs(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validates class and structure of MCMC objects.
}
\details{
To just validate class use \code{\link[chk]{vld_s3_class}()}.
}
\section{Functions}{
\itemize{
\item \code{vld_mcmcarray}: Validate [mcmcarray-object()]

\item \code{vld_mcmcr}: Validate [mcmcr-object()]

\item \code{vld_mcmcrs}: Validate [mcmcrs-object()]
}}

\examples{
#' vld_mcmcarray
vld_mcmcarray(1)

# vld_mcmcr
vld_mcmcr(1)
vld_mcmcr(mcmcr::mcmcr_example)

# vld_mcmcrs
vld_mcmcrs(1)
}
\seealso{
[chk_mcmcr()]
}
