% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.R
\name{zero}
\alias{zero}
\alias{zero.mcarray}
\alias{zero.mcmcarray}
\alias{zero.mcmcr}
\title{Zero MCMC Sample Values}
\usage{
zero(x, ...)

\method{zero}{mcarray}(x, ...)

\method{zero}{mcmcarray}(x, ...)

\method{zero}{mcmcr}(x, pars = NULL, ...)
}
\arguments{
\item{x}{The MCMC object.}

\item{...}{Unused.}

\item{pars}{A character vector (or NULL) of the pars to zero.}
}
\value{
The MCMC
}
\description{
Zeros an MCMC object's sample values.
}
\details{
It is used for removing the effect of a random effect
where the expected value is 0.
}
\section{Methods (by class)}{
\itemize{
\item \code{mcarray}: Zero an mcarray object

\item \code{mcmcarray}: Zero an mcmcarray object

\item \code{mcmcr}: Zero an mcmcr object
}}

\examples{
zero(mcmcr_example, pars = "beta")
}
