% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcarray.R
\name{as.mcarray}
\alias{as.mcarray}
\title{Coerce to an mcarray object}
\usage{
as.mcarray(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to an mcarray object.
}
\examples{
as.mcarray(mcmcr_example$beta)
}
