% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcarray}
\alias{as.mcarray}
\alias{as.mcarray.mcarray}
\alias{as.mcarray.mcmcarray}
\title{Coerce to an mcarray object}
\usage{
as.mcarray(x, ...)

\method{as.mcarray}{mcarray}(x, ...)

\method{as.mcarray}{mcmcarray}(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mcarray object
}
\section{Methods (by class)}{
\itemize{
\item \code{mcarray}: Coerces mcarray

\item \code{mcmcarray}: Coerces mcmcarray
}}

\examples{
as.mcarray(mcmcr_example$beta)
}
