% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcarray}
\alias{as.mcarray}
\alias{as.mcarray.mcmcarray}
\title{Coerce to an mcarray object}
\usage{
as.mcarray(x, ...)

\method{as.mcarray}{mcmcarray}(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to an mcarray object.
}
\section{Methods (by class)}{
\itemize{
\item \code{mcmcarray}: Coerces mcmcarray object to an mcarray object
}}

\examples{
as.mcarray(mcmcr_example$beta)
}
