% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcmc.list}
\alias{as.mcmc.list}
\alias{as.mcmc.list.mcarray}
\alias{as.mcmc.list.mcmcarray}
\alias{as.mcmc.list.mcmc}
\alias{as.mcmc.list.mcmc.list}
\alias{as.mcmc.list.mcmcr}
\title{Coerce to an mcmc.list object}
\usage{
\method{as.mcmc.list}{mcarray}(x, ...)

\method{as.mcmc.list}{mcmcarray}(x, ...)

\method{as.mcmc.list}{mcmc}(x, ...)

\method{as.mcmc.list}{mcmc.list}(x, ...)

\method{as.mcmc.list}{mcmcr}(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mcmc.list object
}
\section{Methods (by class)}{
\itemize{
\item \code{mcarray}: Coerces mcarray

\item \code{mcmcarray}: Coerces mcmcarray

\item \code{mcmc}: Coerces mcmc

\item \code{mcmc.list}: Coerces mcmc.list

\item \code{mcmcr}: Coerces mcmcr
}}

\examples{
as.mcmc.list(mcmcr_example)
}
\seealso{
\code{coda::\link[coda]{as.mcmc.list}}
}
