% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc.R
\name{as.mcmc}
\alias{as.mcmc}
\title{Coerce to an mcmc object}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to an \code{\link[coda]{mcmc}} object.
}
\examples{
as.mcmc(subset(mcmcr_example, chains = 1L))
}
\seealso{
\code{coda::\link[coda]{mcmc}}
}
