% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmcarray.R
\name{as.mcmcarray}
\alias{as.mcmcarray}
\title{Coerce to an mcmcarray object}
\usage{
as.mcmcarray(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to an \code{\link{mcmcarray-object}}.
}
\examples{
as.mcmcarray(as.mcarray(mcmcr_example$beta))
}
