% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcmcarray}
\alias{as.mcmcarray}
\title{Coerce to an mcmcarray object}
\usage{
as.mcmcarray(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mcmcarray object
}
\examples{
as.mcmcarray(list_mcarrays[[1]])
}
