% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.mcmcr}
\alias{as.mcmcr}
\title{Coerce to an mcmcr object}
\usage{
as.mcmcr(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerce to an mcmcr object
}
\examples{
as.mcmcr(coda::as.mcmc.list(mcmcr_example))
}
