% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmcrs.R
\name{as.mcmcrs}
\alias{as.mcmcrs}
\title{Coerce to an mcmcrs object}
\usage{
as.mcmcrs(x, ...)
}
\arguments{
\item{x}{object to coerce.}

\item{...}{Unused.}
}
\description{
Coerces MCMC objects to an \code{\link{mcmcrs-object}}.
}
\examples{
as.mcmcrs(list(mcmcr_example))
}
