% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{as.term}
\alias{as.term}
\title{Coerce object vector to term vector.}
\usage{
as.term(x, ...)
}
\arguments{
\item{x}{The object to coerce}

\item{...}{Unused.}
}
\description{
Coerce object vector to term vector.
}

