% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains.mcmcr}
\alias{bind_chains.mcmcr}
\title{Bind by Chains.}
\usage{
\method{bind_chains}{mcmcr}(x, x2, ...)
}
\arguments{
\item{x}{An object.}

\item{x2}{A second object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The combined object.
}
\description{
Binds two MCMC objects (with the same parameters and iterations) by chains.
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_iterations}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()},
\code{\link[universals]{split_chains}()}
}
