% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-chains.R
\name{bind_chains}
\alias{bind_chains}
\title{Combine MCMC objects by chains.}
\usage{
bind_chains(x, x2, ...)
}
\arguments{
\item{x}{an MCMC object.}

\item{x2}{a second MCMC object}

\item{...}{Unused.}
}
\description{
Combines two MCMC objects (with the same parameters and iterations) by chains.
}
\examples{
bind_chains(mcmcr_example, mcmcr_example)
}
