% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-dimensions.R
\name{bind_dimensions}
\alias{bind_dimensions}
\title{Combines objects by samples}
\usage{
bind_dimensions(x, x2, along = NULL, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{along}{A count (or NULL) indicating the parameter dimension to bind along.}

\item{...}{Unused.}
}
\description{
Combines two mcmc objects by samples.
}
\details{
They must have the same parameters with the same number of chains and iterations.
}
