% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-dimensions-n.R
\name{bind_dimensions_n}
\alias{bind_dimensions_n}
\title{Combines objects by samples}
\usage{
bind_dimensions_n(...)
}
\arguments{
\item{...}{One or more mcmc objects.}
}
\description{
Combines multiple mcmc objects by samples.
}
\details{
They must have the same parameters with the same parameter dimensions and number of chains and iterations.
}
\examples{
bind_dimensions_n(mcmcr_example, mcmcr_example, mcmcr_example)
}
