% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-dimensions-n.R
\name{bind_dimensions_n}
\alias{bind_dimensions_n}
\title{Combine multiple MCMC objects by parameter dimensions}
\usage{
bind_dimensions_n(...)
}
\arguments{
\item{...}{one or more MCMC objects}
}
\description{
Combines multiple MCMC objects (with the same parameters, chains and iterations) by parameter dimensions.
}
\examples{
bind_dimensions_n(mcmcr_example, mcmcr_example, mcmcr_example)
}
\seealso{
\code{\link[universals:bind_chains]{universals::bind_chains()}}

Other bind: 
\code{\link{bind_dimensions}()},
\code{\link{bind_parameters}()}
}
\concept{bind}
