% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-parameters.R
\name{bind_parameters}
\alias{bind_parameters}
\title{Combines objects by parameters}
\usage{
bind_parameters(x, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{...}{Unused.}
}
\description{
Combines an mcmc object(s) by parameters.
}
