% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-parameters.R
\name{bind_parameters}
\alias{bind_parameters}
\title{Combine two MCMC object by parameters}
\usage{
bind_parameters(x, x2, ...)
}
\arguments{
\item{x}{an MCMC object}

\item{x2}{a second MCMC object.}

\item{...}{Unused.}
}
\description{
Combines two MCMC objects (with the same chains and iterations) by their parameters.
}
\examples{
bind_parameters(
  subset(mcmcr_example, pars = "sigma"),
  subset(mcmcr_example, pars = "beta")
)
}
\seealso{
\code{\link[universals:bind_chains]{universals::bind_chains()}}

Other bind: 
\code{\link{bind_dimensions_n}()},
\code{\link{bind_dimensions}()}
}
\concept{bind}
