% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind-samples.R
\name{bind_samples}
\alias{bind_samples}
\title{Combines objects by samples}
\usage{
bind_samples(x, x2, ...)
}
\arguments{
\item{x}{an mcmc object.}

\item{x2}{a second mcmc object.}

\item{...}{Unused.}
}
\description{
Combines two mcmc objects by samples
}
