% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.mcmcarray}
\alias{coef.mcmcarray}
\title{Coef TMB Analyses}
\usage{
\method{coef}{mcmcarray}(object, conf_level = 0.95, ...)
}
\arguments{
\item{object}{The list of tmb_analysis objects.}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{...}{Not used.}
}
\value{
A tidy tibble of the coeffcient terms with the model averaged estimate, the
Akaike's weight and the proportion of models including the term.
}
\description{
Coefficients for an mcmcarray object.
}
