% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-dimensions.R
\name{combine_dimensions}
\alias{combine_dimensions}
\title{Combine samples by dimensions}
\usage{
combine_dimensions(x, fun = mean, along = NULL, ...)
}
\arguments{
\item{x}{An MCMC object}

\item{fun}{The function to use when combining dimensions}

\item{along}{A positive integer (or NULL) indicating the parameter dimension(s) to bind along.}

\item{...}{Unused.}
}
\value{
The MCMC object with reduced dimensions.
}
\description{
Combines MCMC object samples by dimensions along \code{along} using \code{fun}.
}
\examples{
combine_dimensions(mcmcr_example$alpha)
}
\seealso{
Other combine: 
\code{\link{combine_samples_n}()},
\code{\link{combine_samples}()}
}
\concept{combine}
