% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-samples.R
\name{combine_samples}
\alias{combine_samples}
\title{Combines object samples.}
\usage{
combine_samples(x, x2, fun = mean, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{x2}{A second mcmc object.}

\item{fun}{A function.}

\item{...}{Unused.}
}
\description{
Combines mcmc objects by values using fun.
}
