% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-samples-n.R
\name{combine_samples_n}
\alias{combine_samples_n}
\title{Combine MCMC Samples of multiple objects}
\usage{
combine_samples_n(x, ..., fun = mean)
}
\arguments{
\item{x}{An MCMC object (or a list of mcmc objects).}

\item{...}{Additional MCMC objects.}

\item{fun}{A function.}
}
\description{
Combines samples of multiple MCMC objects
(with the same parameters, chains and iterations) using a function.
}
\examples{
combine_samples_n(mcmcr_example, mcmcr_example, mcmcr_example, fun = sum)
}
