% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-samples-n.R
\name{combine_samples_n}
\alias{combine_samples_n}
\title{Combines multiple object samples.}
\usage{
combine_samples_n(x, ..., fun = mean)
}
\arguments{
\item{x}{An mcmc object (or a list of mcmc objects).}

\item{...}{Additional mcmc objects (or additional arguments).}

\item{fun}{A function.}
}
\description{
Combines mcmc objects by values using fun.
}
\examples{
combine_samples_n(mcmcr_example, mcmcr_example, mcmcr_example, fun = sum)
}
