% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-values.R
\name{combine_values_list}
\alias{combine_values_list}
\title{Combine list of mcmc samples}
\usage{
combine_values_list(x, fun = mean)
}
\arguments{
\item{x}{An mcmc object.}

\item{fun}{A function.}
}
\description{
Combine list of mcmc samples
}
