% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converged.R
\name{converged.default}
\alias{converged.default}
\title{Converged}
\usage{
\method{converged}{default}(x, rhat = 1.1, esr = 0.33, by = "all", as_df = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{rhat}{The maximum rhat value.}

\item{esr}{The minimum effective sampling rate.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{...}{Other arguments passed to methods.}
}
\value{
A logical scalar indicating whether the object has converged.
}
\description{
Tests whether an object has converged.
}
\examples{
converged(mcmcr_example)
}
\seealso{
Other {convergence}: 
\code{\link[universals]{esr}()},
\code{\link[universals]{rhat}()}
}
