% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{converged}
\alias{converged}
\title{Is Converged}
\usage{
converged(x, ...)
}
\arguments{
\item{x}{The object to test.}

\item{...}{Unused}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x has converged.
}
