% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{derive}
\usage{
derive(object, ...)
}
\arguments{
\item{object}{The mcmc object.}

\item{...}{Unused.}
}
\value{
An mcmc object of the derived parameters.
}
\description{
Calculate dervived parameters.
}
