% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr.R
\name{esr}
\alias{esr}
\title{Effective Sampling Rate}
\usage{
esr(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the esr value.
}
\description{
Calculates the effective sampling rate (esr) based on the formula
\deqn{\frac{1}{1 + 2 \sum_{k = 1}^\infty\rho_k(\theta)}}
in the Handbook for Markov Chain Monte Carlo.
}
\details{
The infinite sum is truncated at lag \eqn{k} when \eqn{\rho_k(θ) < 0}.
}
\references{
Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.
}
