% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esr.R
\name{esr}
\alias{esr}
\title{Effective Sampling Rate}
\usage{
esr(x, ...)
}
\arguments{
\item{x}{A vector, matrix (each row is a chain) or mcmc object.}

\item{...}{Unused.}
}
\value{
The esr value(s).
}
\description{
Calculates the effective sampling rate (\code{esr}) based on the formula
\deqn{\frac{1}{1 + 2 \sum_{k = 1}^\infty\rho_k(\theta)}}
in Brooks et al. (2011).
The infinite sum is truncated at
lag \eqn{k} when \eqn{\rho_{k+1}(\theta) < 0}.
}
\examples{
esr(rnorm(1000))
}
\references{
Brooks, S., Gelman, A., Jones, G.L., and Meng, X.-L. (Editors). 2011. Handbook for Markov Chain Monte Carlo. Taylor & Francis, Boca Raton.

Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A., and
Rubin, D.B. 2014. Bayesian data analysis. In Third edition. CRC Press,
Boca Raton.
}
