% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{Effective Sample Size}
\usage{
ess(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the ess value.
}
\description{
Effective Sample Size
}
\examples{
ess(mcmcr_example)
}
