% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ess}
\alias{ess}
\title{Effective Sample Size}
\usage{
ess(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A count of the effective total number of samples.
}
\description{
The effective number of MCMC samples.
}
