% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ess}
\alias{ess}
\title{Effective Sample Size}
\usage{
ess(x, by = "all", as_df = FALSE)
}
\arguments{
\item{x}{An mcmc object.}

\item{by}{A string indicating whether to determine by "term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the results as a (tbl) data frame.}
}
\description{
Effective Sample Size
}
\examples{
ess(mcmcr_example)
}
