% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\title{Get Estimates}
\usage{
estimates(object, ...)
}
\arguments{
\item{object}{The object.}

\item{...}{Not used.}
}
\description{
Gets the estimates for an object as a named list.
}
