% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_.mcmcr_data}
\alias{filter_.mcmcr_data}
\title{Filter mcmcr data}
\usage{
\method{filter_}{mcmcr_data}(.data, ..., .dots)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Pair/values of expressions coercible to lazy objects.}

\item{.dots}{Pair/values of expressions coercible to lazy objects.}
}
\description{
Filter mcmcr data
}
