% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_.mcmcr_data}
\alias{filter_.mcmcr_data}
\title{Filter mcmcr data}
\usage{
\method{filter_}{mcmcr_data}(.data, ..., .dots)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Logical predicates. Multiple conditions are combined with \code{&}.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Filter mcmcr data
}
