% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{group_by_.mcmcr_data}
\alias{group_by_.mcmcr_data}
\title{Group mcmcr data}
\usage{
group_by_.mcmcr_data(.x, ..., .dots, add = FALSE)
}
\arguments{
\item{.x}{An mcmcr_data object}

\item{...}{variables to group by. All tbls accept variable names,
some will also accept functions of variables. Duplicated groups
will be silently dropped.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}

\item{add}{By default, when \code{add = FALSE}, \code{group_by} will
override existing groups. To instead add to the existing groups,
use \code{add = TRUE}}
}
\description{
Group mcmcr data
}
