% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcarray}
\alias{is.mcarray}
\title{Is mcarray Object}
\usage{
is.mcarray(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an object is an mcarray.
}
\examples{
is.mcarray(mcmcr_example)
}
