% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcarray}
\alias{is.mcarray}
\title{Is mcarray object}
\usage{
is.mcarray(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an object is an mcarray.
}
\examples{
is.mcarray(mcmcr_example)
}
\seealso{
Other is: 
\code{\link{is.mcmcarray}()},
\code{\link{is.mcmcrs}()},
\code{\link{is.mcmcr}()}
}
\concept{is}
