% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmcarray}
\alias{is.mcmcarray}
\title{Is an mcmcrarray object}
\usage{
is.mcmcarray(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mcmcarray'
}
