% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.mcmcr}
\alias{is.mcmcr}
\title{Is mcmcr Object}
\usage{
is.mcmcr(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an object is an \code{\link[=mcmcr-object]{mcmcr-object()}}.
}
\examples{
is.mcmcr(mcmcr_example)
}
