% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{is.term}
\alias{is.term}
\title{Test whether an object is a term.}
\usage{
is.term(x, ...)
}
\arguments{
\item{x}{The object to test.}

\item{...}{Unused.}
}
\description{
Test whether an object is a term.
}

