% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-aperm.R
\name{mcmc_aperm}
\alias{mcmc_aperm}
\title{MCMC Object Transposition}
\usage{
mcmc_aperm(x, perm, ...)
}
\arguments{
\item{x}{The MCMC object to transpose.}

\item{perm}{A integer vector of the new order for the parameter dimensions.
Missing parameter dimensions are added on the end.
If perm = NULL (the default) the parameter dimensions are reversed.}

\item{...}{Unused}
}
\value{
The modified MCMC object
}
\description{
Transpose an MCMC object by permuting its parameter dimensions.
}
