% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-map.R
\name{mcmc_map}
\alias{mcmc_map}
\title{MCMC map}
\usage{
mcmc_map(.x, .f, ...)
}
\arguments{
\item{.x}{The mcmc object.}

\item{.f}{The function to use}

\item{...}{Additional arguments passed to .f.}
}
\value{
The mapped mcmc object.
}
\description{
Map mcmc values
}
