% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-map.R
\name{mcmc_map}
\alias{mcmc_map}
\title{MCMC Map}
\usage{
mcmc_map(.x, .f, .by = 1:npdims(.x), ...)
}
\arguments{
\item{.x}{An MCMC object}

\item{.f}{The function to use}

\item{.by}{A positive integer vector of the dimensions to apply the function over.}

\item{...}{Additional arguments passed to .f.}
}
\value{
The updated MCMC object.
}
\description{
Adjust the sample values of an MCMC object using a function.
}
\examples{
mcmc_map(mcmcr_example$beta, exp)
}
