% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcarray-object.R
\name{mcmcarray-object}
\alias{mcmcarray-object}
\alias{mcmcarray_object}
\title{mcmcarray}
\description{
An \code{mcmcarray} object is an an array where the
first dimension is the chains, the second dimension is the iterations
and the subsequent dimensions represent the dimensionality of the parameter.
The name \code{mcmcarray} reflects the fact that the MCMC dimensions,
ie the chains and iterations, precede the parameter dimensions.
}
\examples{
mcmcr_example$beta
}
