% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-package.R
\docType{package}
\name{mcmcr-package}
\alias{mcmcr}
\alias{mcmcr-package}
\title{mcmcr: Store, Manipulate and Summarise MCMC Samples}
\description{
Defines S3 classes to store Monte Carlo Markov Chain (MCMC) samples and
provides functions to manipulate and summarise MCMC samples.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
