% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-package.R
\docType{package}
\name{mcmcr-package}
\alias{mcmcr}
\alias{mcmcr-package}
\title{mcmcr: Manipulate MCMC Samples}
\description{
Manipulates Monte Carlo Markov Chain (MCMC) samples.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

}
\keyword{internal}
