% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-package.R
\docType{package}
\name{mcmcr-package}
\alias{mcmcr}
\alias{mcmcr-package}
\title{mcmcr: Manipulate MCMC Samples}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and classes to store, manipulate and summarise Monte Carlo Markov Chain (MCMC) samples. For more information see Brooks et al. (2011) <isbn:978-1-4200-7941-8>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/mcmr}
  \item \url{https://poissonconsulting.github.io/mcmcr/}
  \item Report bugs at \url{https://github.com/poissonconsulting/mcmcr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID}) [contributor]
  \item Nadine Hussein (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
  \item Ayla Pearson (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
