% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcr-package.R
\docType{package}
\name{mcmcr-package}
\alias{mcmcr}
\alias{mcmcr-package}
\title{mcmcr: Manipulate MCMC Samples}
\description{
Functions and classes to store, manipulate and summarise 
   Monte Carlo Markov Chain (MCMC) samples. For more information see 
   Brooks et al. (2011) <isbn:978-1-4200-7941-8>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/mcmcr}
  \item Report bugs at \url{https://github.com/poissonconsulting/mcmcr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
