% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcr_example}
\alias{mcmcr_example}
\title{Example mcmcr Object}
\format{An object of class \code{mcmcr} of length 3.}
\usage{
mcmcr_example
}
\description{
An example mcmcr object based on coda::line.
}
\details{
It consists of three chains of 200 iterations of three parameters.
}
\keyword{datasets}
