% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmcr_example}
\alias{mcmcr_example}
\title{An Example mcmcr Object}
\format{An object of class \code{mcmcr} of length 3.}
\usage{
mcmcr_example
}
\description{
An example \code{\link{mcmcr-object}}
derived from \code{coda::\link[coda]{line}}.
}
\examples{
mcmcr_example
}
\keyword{datasets}
