% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchains.R
\name{nchains}
\alias{nchains}
\title{Number of MCMC chains}
\usage{
nchains(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count indicating the number of MCMC chains in object.
}
\description{
Number of MCMC chains
}
\examples{
nchains(mcmcr_example)
}
