% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters}
\alias{niters}
\title{Number of MCMC samples}
\usage{
niters(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count indicating the number of MCMC iterations.
}
\description{
Gets the number of MCMC iterations (in a chain).
}
\examples{
niters(mcmcr_example)
}
