% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niters.R
\name{niters}
\alias{niters}
\title{Number of MCMC samples in a single chain}
\usage{
niters(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A count indicating the number of MCMC samples in a chain.
}
\description{
Number of MCMC samples in a single chain
}
\examples{
niters(mcmcr_example)
}
