% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{npars.mcmcr}
\alias{npars.mcmcr}
\title{Number of parameters}
\usage{
\method{npars}{mcmcr}(x, scalar_only = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{scalar_only}{A flag indicating whether to only count scalar parameters.}

\item{...}{Not used.}
}
\description{
Number of parameters
}
