% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{npars}
\alias{npars}
\title{Number of parameters}
\usage{
npars(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A count of the number of parameters.
}
\description{
Gets the number of parameter names for an object.
}

