% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\title{Number of Parameters}
\usage{
npars(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of parameters for an object.
}
\examples{
npars(mcmcr_example)
}
