% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nsims}
\alias{nsims}
\title{Number of MCMC Simulations}
\usage{
nsims(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A count of the total number of simulations.
}
\description{
The product of the number of chains and number of iterations in each chain.
}
\examples{
nsims(mcmcr_example)
}
