% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{nterms}
\alias{nterms}
\title{Number of terms.}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{The object to calculate the number of terms for.}

\item{...}{Not used.}
}
\value{
A count of the number of terms.
}
\description{
Number of terms.
}

