% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of terms}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms for an object.
}
\examples{
nterms(mcmcr_example)
}
