% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters<-}
\alias{parameters<-}
\title{Parameters}
\usage{
parameters(x) <- value
}
\arguments{
\item{x}{The object.}

\item{value}{A character vector of the parameter names.}
}
\description{
Sets the parameter names for an object.
}
