% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R, R/scalar_parameters.R
\name{parameters}
\alias{parameters}
\alias{parameters}
\title{Parameters}
\usage{
parameters(x, ...)

parameters(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Not used.}

\item{x}{The object.}

\item{...}{Not used.}
}
\value{
A character vector of the parameter names.

A character vector of the parameter names.
}
\description{
Gets the parameter names for an object.

Gets the parameter names for an object.
}
