% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to only include scalar terms.}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{nas}{A flag specifying whether to also fill missing values.}

\item{nthin}{A positive integer of the thinning rate.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{bound}{flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.}

\item{rhat}{The maximum rhat value.}

\item{esr}{The minimum effective sampling rate.}

\item{...}{Unused.}
}
\description{
Parameter Descriptions
}
