% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter descriptions}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{nas}{A flag specifying whether to also fill missing values.}

\item{nthin}{A positive integer of the thinning rate.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the results as a
data frame versus a named list.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{bound}{flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.}

\item{rhat}{The maximum rhat value.}

\item{esr}{The minimum effective sampling rate.}

\item{na_rm}{A flag specifying whether to ignore missing values.}

\item{parameters}{A character vector (or NULL) of the parameters to subset by.}

\item{iterations}{An integer vector (or NULL) of the iterations to subset by.}

\item{...}{Unused.}

\item{x2}{a second MCMC object.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}

\item{sorted}{A flag specifying whether the parameters must be sorted.}

\item{object}{The MCMC object to get the coefficients for}

\item{conf_level}{A number specifying the confidence level. By default 0.95.}

\item{estimate}{The function to use to calculate the estimate.}

\item{simplify}{A flag specifying whether to return just the estimate, lower, upper and svalue.}

\item{perm}{A integer vector of the new order for the parameter dimensions.}

\item{pars}{A character vector (or NULL) of the pars to zero.}

\item{name}{A string specifying the parameter name.}
}
\description{
Parameter descriptions
}
