% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NULL), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{nas}{A flag specifying whether to also fill missing values.}

\item{nthin}{A positive integer of the thinning rate.}

\item{by}{A string indicating whether to determine by
"term", "parameter" or "all".}

\item{as_df}{A flag indicating whether to return the values as a
data frame versus a named list.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{bound}{flag specifying whether to bind mcmcrs objects by their chains before calculating rhat.}

\item{rhat}{The maximum rhat value.}

\item{esr}{The minimum effective sampling rate.}

\item{na_rm}{A flag specifying whether to ignore missing values.}

\item{parameters}{A character vector (or NULL) of the parameters to subset by.}

\item{iterations}{An integer vector (or NULL) of the iterations to subset by.}

\item{...}{Unused.}
}
\description{
Parameter Descriptions
}
