% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to only include scalar terms.}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{nas}{A flag specifying whether to also fill missing values.}
}
\description{
Parameter Descriptions
}
