% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.mcmc.list}
\alias{pars.mcmc.list}
\title{Get Parameters}
\usage{
\method{pars}{mcmc.list}(x, scalar = NA, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to only include scalar terms.}

\item{terms}{A logical scalar specifying whether to provide the parameters for each term.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets character vector of parameter names.
}
\seealso{
Other {parameters}: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars_terms}()},
\code{\link[universals]{set_pars}()}
}
