% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims.mcmcr}
\alias{pdims.mcmcr}
\title{Parameter Dimensions}
\usage{
\method{pdims}{mcmcr}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of integer vectors of the dimensions of each parameter.
}
\description{
Gets the dimension of each parameter of an object.
}
\seealso{
Other {dimensions}: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{tindex}()}
}
