% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mcmcr}
\alias{predict.mcmcr}
\title{Predict}
\usage{
\method{predict}{mcmcr}(object, expr, values = list(), monitor = ".*", ...)
}
\arguments{
\item{object}{The mcmcr object.}

\item{expr}{A string of the R expression to evaluate.}

\item{values}{A lists of the values to use.}

\item{monitor}{A regular expression specifying the new variables to monitor.}

\item{...}{Unused.}
}
\value{
An mcmcr object of the monitored parameters.
}
\description{
Calculate predictions.
}

