% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term.R, R/niters.R, R/nterms.R, R/pdims.R,
%   R/split-chains.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.term}
\alias{niters}
\alias{nterms}
\alias{pdims}
\alias{split_chains}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{term}{\code{\link[term]{as.term}}}

  \item{universals}{\code{\link[universals]{niters}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pdims}}, \code{\link[universals]{split_chains}}}
}}

