% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc.R, R/as-nlist.R, R/as-nlists.R,
%   R/as-term.R, R/bind-chains.R, R/collapse-chains.R, R/converged.R, R/esr.R,
%   R/estimates.R, R/export.R, R/exports.R, R/fill-all.R, R/nchains.R,
%   R/niters.R, R/npars.R, R/npdims.R, R/nterms.R, R/pars.R, R/pdims.R,
%   R/rhat.R, R/set-pars.R, R/split-chains.R, R/thin.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.mcmc}
\alias{as_nlist}
\alias{as_nlists}
\alias{as.term}
\alias{bind_chains}
\alias{collapse_chains}
\alias{converged}
\alias{esr}
\alias{estimates}
\alias{median}
\alias{zscore}
\alias{lower}
\alias{upper}
\alias{svalue}
\alias{nlist}
\alias{nlists}
\alias{pvalue}
\alias{fill_all}
\alias{nchains}
\alias{niters}
\alias{npars}
\alias{npdims}
\alias{nterms}
\alias{pars}
\alias{pdims}
\alias{rhat}
\alias{set_pars}
\alias{split_chains}
\alias{thin}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda]{as.mcmc}}, \code{\link[coda]{thin}}}

  \item{extras}{\code{\link[extras]{fill_all}}, \code{\link[extras]{lower}}, \code{\link[extras]{pvalue}}, \code{\link[extras]{svalue}}, \code{\link[extras]{upper}}, \code{\link[extras]{zscore}}}

  \item{generics}{\code{\link[generics]{tidy}}}

  \item{nlist}{\code{\link[nlist]{as_nlist}}, \code{\link[nlist]{as_nlists}}, \code{\link[nlist]{nlist}}, \code{\link[nlist]{nlists}}}

  \item{stats}{\code{\link[stats]{median}}}

  \item{term}{\code{\link[term]{as.term}}}

  \item{universals}{\code{\link[universals]{bind_chains}}, \code{\link[universals]{collapse_chains}}, \code{\link[universals]{converged}}, \code{\link[universals]{esr}}, \code{\link[universals]{estimates}}, \code{\link[universals]{nchains}}, \code{\link[universals]{niters}}, \code{\link[universals]{npars}}, \code{\link[universals]{npdims}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pars}}, \code{\link[universals]{pdims}}, \code{\link[universals]{rhat}}, \code{\link[universals]{set_pars}}, \code{\link[universals]{split_chains}}}
}}

