% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-mcmc.R, R/as-nlist.R, R/as-term.R,
%   R/bind-chains.R, R/collapse-chains.R, R/nchains.R, R/niters.R, R/npars.R,
%   R/nterms.R, R/pars.R, R/pdims.R, R/set-pars.R, R/split-chains.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.mcmc}
\alias{as.nlist}
\alias{as.term}
\alias{bind_chains}
\alias{collapse_chains}
\alias{niters}
\alias{npars}
\alias{nterms}
\alias{pars}
\alias{pdims}
\alias{set_pars}
\alias{split_chains}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coda}{\code{\link[coda]{as.mcmc}}}

  \item{nlist}{\code{\link[nlist]{as.nlist}}}

  \item{term}{\code{\link[term]{as.term}}}

  \item{universals}{\code{\link[universals]{bind_chains}}, \code{\link[universals]{collapse_chains}}, \code{\link[universals]{niters}}, \code{\link[universals]{niters}}, \code{\link[universals]{npars}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pars}}, \code{\link[universals]{pdims}}, \code{\link[universals]{set_pars}}, \code{\link[universals]{split_chains}}}
}}

