% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat}
\alias{rhat}
\title{Convergence}
\usage{
rhat(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
A number of the rhat value.
}
\description{
Get R-hat value(s) for JAGS objects
}

