% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhat.R
\name{rhat}
\alias{rhat}
\title{Rhat}
\usage{
rhat(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\value{
The rhat value(s).
}
\description{
Calculates the R-hat values
Gets uncorrected, untransformed, univariate split R-hat value(s).
}
