% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-pars.R
\name{set_pars.mcmcrs}
\alias{set_pars.mcmcrs}
\title{Set Parameters}
\usage{
\method{set_pars}{mcmcrs}(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A character vector of the new parameter names.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Sets an object's parameter names.
}
\details{
\code{value} must be a unique character vector of the same length as the object's parameters.
}
\examples{
set_pars.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to set_pars for an object of class 'foobar'
}
}
\seealso{
Other parameters: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars}()}
}
