% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_.mcmcr_data}
\alias{slice_.mcmcr_data}
\title{Slice mcmcr data}
\usage{
\method{slice_}{mcmcr_data}(.data, ..., .dots)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Integer row values}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Slice mcmcr data
}
