% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains.mcmcr}
\alias{split_chains.mcmcr}
\title{Split Chains}
\usage{
\method{split_chains}{mcmcr}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Splits each of an MCMC object's chains in half
to double the number chains
and halve the number of iterations.
}
\examples{
split_chains.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to split_chains for an object of class 'foobar'
}
}
\seealso{
Other MCMC manipulations: 
\code{\link[universals]{bind_chains}()},
\code{\link[universals]{collapse_chains}()},
\code{\link[universals]{estimates}()}
}
