% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-chains.R
\name{split_chains}
\alias{split_chains}
\title{Split Chains}
\usage{
split_chains(x, ...)
}
\arguments{
\item{x}{An mcmc object.}

\item{...}{Unused.}
}
\description{
Splits each chain into two chains.
}
\examples{
split_chains(mcmcr_example)
}
