% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_.mcmcr_data}
\alias{summarise_.mcmcr_data}
\title{Summarise mcmcr data}
\usage{
\method{summarise_}{mcmcr_data}(.data, ..., .dots)
}
\arguments{
\item{.data}{An mcmcr_data object}

\item{...}{Name-value pairs of summary functions like \code{\link{min}()},
\code{\link{mean}()}, \code{\link{max}()} etc.}

\item{.dots}{Used to work around non-standard evaluation. See
\code{vignette("nse")} for details.}
}
\description{
Summarise mcmcr data
}
