% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thin}
\alias{thin}
\title{Thin}
\usage{
thin(x, nthin = 1L, ...)
}
\arguments{
\item{x}{The object to thin}

\item{nthin}{A count of the thinning interval}

\item{...}{Not used.}
}
\description{
Thin
}
\examples{
thin(mcmcr::mcmcr_example)
}
