% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_mcmcr}
\alias{vld_mcmcr}
\alias{vld_mcmcarray}
\alias{vld_mcmcrs}
\title{Validate MCMC Objects}
\usage{
vld_mcmcarray(x)

vld_mcmcr(x)

vld_mcmcrs(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validates class and structure of MCMC objects.
}
\details{
To simply validate class use for example \code{\link[chk]{vld_s3_class}(x, "mcmcr")}.
}
\section{Functions}{
\itemize{
\item \code{vld_mcmcarray}: Validate mcmcr Object

\item \code{vld_mcmcr}: Validate mcmcr Object

\item \code{vld_mcmcrs}: Validate mcmcrs Object
}}

\examples{
#' vld_natomic
vld_mcmcarray(1)

# vld_mcmcr
vld_mcmcr(1)
vld_mcmcr(as.mcmcr(list(x = 1)))

# vld_mcmcrs
vld_mcmcrs(1)
}
