% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_mcmcr}
\alias{vld_mcmcr}
\alias{vld_mcmcarray}
\title{Validate MCMC Objects}
\usage{
vld_mcmcarray(x)

vld_mcmcr(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
A flag indicating whether the object was validated.
}
\description{
Validates structure of MCMC objects.
}
\details{
To validate the class use \code{\link[chk]{vld_s3_class}(x, "class")}.
}
\section{Functions}{
\itemize{
\item \code{vld_mcmcarray}: Validate [mcmcarray-object()]

\item \code{vld_mcmcr}: Validate [mcmcr-object()]
}}

\examples{
#' vld_mcmcarray
vld_mcmcarray(1)

# vld_mcmcr
vld_mcmcr(1)
vld_mcmcr(as.mcmcr(list(x = 1)))
}
\seealso{
[chk_mcmcr()]
}
